# docs/contracts/openapi-v1-runs.yaml
openapi: 3.0.3
info:
  title: Autonoma Runs API
  version: 1.0.0
paths:
  /v1/runs:
    post:
      summary: Create a run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [intent, environment]
              additionalProperties: false
              properties:
                intent:
                  type: string
                  minLength: 1
                environment:
                  type: string
                  enum: [dev, stage, prod]
                metadata:
                  type: object
      responses:
        "202":
          description: Accepted
  /v1/runs/{run_id}:
    get:
      summary: Get run
      parameters:
        - in: path
          name: run_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Run summary
...